refactor: restructure dao pallet #735
